refactor: extract data and controls into separate components
