+++ components/InteractiveOptionsChart.tsx
@@ -1,288 +1,21 @@
 import React, { useEffect, useMemo, useState } from 'react';
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from './ui/card';
+import { Card, CardContent, CardHeader, CardTitle } from './ui/card';
 import { Button } from './ui/button';
-import { Badge } from './ui/badge';
-import { Slider } from './ui/slider';
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from './ui/tooltip';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from './ui/tabs';
-import { ChevronRight, Activity, Zap } from 'lucide-react';
-import { motion, AnimatePresence } from 'motion/react';
-import {
-  ResponsiveContainer,
-  LineChart,
-  Line,
-  CartesianGrid,
-  XAxis,
-  YAxis,
-  Tooltip as RechartsTooltip,
-  ReferenceLine,
-  ReferenceArea,
-} from 'recharts';
+import { Zap } from 'lucide-react';
 import { GreeksExplainer, OptionsData } from './GreeksExplainer';
-import { StrategyVisualizer } from './StrategyVisualizer';
+import { StrategyCard } from './StrategyCard';
+import { OptionsParameters } from './OptionsParameters';
+import {
+  optionsStrategies,
+  OptionsStrategy,
+  PayoffPoint,
+} from '../lib/strategies';
+import {
+  validateParameters,
+  safePayoffCalculation,
+} from '../lib/optionsUtils';
 
-interface PayoffPoint {
-  stockPrice: number;
-  profitLoss: number;
-}
-
-interface OptionsStrategy {
-  id: string;
-  title: string;
-  description: string;
-  category: string;
-  complexity: string;
-  riskLevel: string;
-  maxProfit: string;
-  maxLoss: string;
-  breakeven: string;
-  whenToUse: string[];
-  calculatePayoff: (stockPrice: number, options: OptionsData) => number;
-  legs: { action: string; type: string; strike?: number; premium?: number }[];
-}
-
-const optionsStrategies: OptionsStrategy[] = [
-  {
-    id: 'long-call',
-    title: 'Long Call',
-    description:
-      'Buy a call option expecting the stock price to rise significantly above the strike price.',
-    category: 'Bullish',
-    complexity: 'Basic',
-    riskLevel: 'Low',
-    maxProfit: 'Unlimited',
-    maxLoss: 'Premium Paid',
-    breakeven: 'Strike Price + Premium',
-    whenToUse: [
-      'You expect the stock to rise significantly',
-      'Earnings announcement approaching with positive expectations',
-      'Technical breakout patterns suggesting upward momentum',
-      'Low cost way to participate in upside potential',
-    ],
-    calculatePayoff: (s, o) => Math.max(s - o.strikePrice, 0) - o.premium,
-    legs: [{ action: 'Buy', type: 'Call' }],
-  },
-  {
-    id: 'long-put',
-    title: 'Long Put',
-    description:
-      'Buy a put option expecting the stock price to fall significantly below the strike price.',
-    category: 'Bearish',
-    complexity: 'Basic',
-    riskLevel: 'Low',
-    maxProfit: 'Strike Price - Premium',
-    maxLoss: 'Premium Paid',
-    breakeven: 'Strike Price - Premium',
-    whenToUse: [
-      'You expect the stock to decline significantly',
-      'Negative news or poor earnings outlook for the company',
-      'Hedging against a long position in the underlying stock',
-      'Low cost way to speculate on downside movement',
-    ],
-    calculatePayoff: (s, o) => Math.max(o.strikePrice - s, 0) - o.premium,
-    legs: [{ action: 'Buy', type: 'Put' }],
-  },
-  {
-    id: 'covered-call',
-    title: 'Covered Call',
-    description:
-      'Sell a call option while holding the underlying stock to generate income and cap potential upside.',
-    category: 'Neutral',
-    complexity: 'Basic',
-    riskLevel: 'Medium',
-    maxProfit: 'Strike Price - Stock Cost + Premium',
-    maxLoss: 'Stock Cost - Premium',
-    breakeven: 'Stock Cost - Premium',
-    whenToUse: [
-      'You believe the stock will trade sideways',
-      'You want to generate income from a long stock position',
-      'You are willing to sell your shares at the strike price',
-      'You expect moderate price appreciation but want some downside protection',
-    ],
-    calculatePayoff: (s, o) => {
-      const stockPayoff = s - o.currentPrice;
-      const shortCall = -Math.max(s - o.strikePrice, 0) + o.premium;
-      return stockPayoff + shortCall;
-    },
-    legs: [
-      { action: 'Sell', type: 'Call' },
-      { action: 'Buy', type: 'Stock' },
-    ],
-  },
-  {
-    id: 'protective-put',
-    title: 'Protective Put',
-    description:
-      'Buy a put option to protect a long stock position from downside risk while maintaining upside potential.',
-    category: 'Neutral',
-    complexity: 'Basic',
-    riskLevel: 'Low',
-    maxProfit: 'Unlimited',
-    maxLoss: 'Stock Cost + Premium - Strike Price',
-    breakeven: 'Stock Cost + Premium',
-    whenToUse: [
-      'You own the stock and want downside protection',
-      'Volatile market conditions with uncertain outlook',
-      'Earnings announcements or macro events could cause large drops',
-      'Insurance against a decline while retaining upside exposure',
-    ],
-    calculatePayoff: (s, o) => {
-      const stockPayoff = s - o.currentPrice;
-      const longPut = Math.max(o.strikePrice - s, 0) - o.premium;
-      return stockPayoff + longPut;
-    },
-    legs: [
-      { action: 'Buy', type: 'Stock' },
-      { action: 'Buy', type: 'Put' },
-    ],
-  },
-];
-
-const quickPresets: Record<string, OptionsData> = {
-  ATMOption: {
-    strikePrice: 100,
-    currentPrice: 100,
-    premium: 3,
-    daysToExpiry: 30,
-    impliedVolatility: 20,
-    interestRate: 5,
-    dividendYield: 2,
-  },
-  OTMCall: {
-    strikePrice: 105,
-    currentPrice: 100,
-    premium: 2,
-    daysToExpiry: 30,
-    impliedVolatility: 25,
-    interestRate: 5,
-    dividendYield: 2,
-  },
-  OTMPut: {
-    strikePrice: 95,
-    currentPrice: 100,
-    premium: 2,
-    daysToExpiry: 30,
-    impliedVolatility: 25,
-    interestRate: 5,
-    dividendYield: 2,
-  },
-  HighVol: {
-    strikePrice: 100,
-    currentPrice: 100,
-    premium: 8,
-    daysToExpiry: 7,
-    impliedVolatility: 50,
-    interestRate: 5,
-    dividendYield: 1,
-  },
-};
-
-type TooltipInfo = { title: string; content: string };
-
-const parameterTooltips: Record<keyof OptionsData, TooltipInfo> = {
-  currentPrice: {
-    title: 'Current Stock Price (S)',
-    content:
-      'The current market price of the underlying stock. This determines option moneyness.',
-  },
-  strikePrice: {
-    title: 'Strike Price (K)',
-    content: 'The exercise price of the option contract.',
-  },
-  premium: {
-    title: 'Option Premium',
-    content:
-      'The price paid for the option contract. For short strategies this is the credit received.',
-  },
-  daysToExpiry: {
-    title: 'Days to Expiry (T)',
-    content:
-      'Number of days until the option expires. Shorter durations increase time decay (Theta).',
-  },
-  impliedVolatility: {
-    title: 'Implied Volatility (IV)',
-    content:
-      'Expected volatility of the underlying over the life of the option. Higher IV increases option premiums.',
-  },
-  interestRate: {
-    title: 'Risk-free Interest Rate (r)',
-    content:
-      'Annualized interest rate used in option pricing models. Higher rates generally raise call values and lower put values.',
-  },
-  dividendYield: {
-    title: 'Dividend Yield (q)',
-    content:
-      'Expected annual dividend yield of the underlying stock. Dividends decrease call values and increase put values.',
-  },
-};
-
-const getCategoryColor = (c: string) =>
-  c === 'Bullish'
-    ? 'bg-green-50 text-green-700 border-green-200'
-    : c === 'Bearish'
-    ? 'bg-red-50 text-red-700 border-red-200'
-    : c === 'Neutral'
-    ? 'bg-blue-50 text-blue-700 border-blue-200'
-    : c === 'Volatility'
-    ? 'bg-purple-50 text-purple-700 border-purple-200'
-    : 'bg-blue-50 text-blue-700 border-blue-200';
-
-const getComplexityColor = (c: string) =>
-  c === 'Basic'
-    ? 'bg-green-50 text-green-700 border-green-200'
-    : c === 'Intermediate'
-    ? 'bg-yellow-50 text-yellow-700 border-yellow-200'
-    : c === 'Advanced'
-    ? 'bg-red-50 text-red-700 border-red-200'
-    : 'bg-gray-50 text-gray-700 border-gray-200';
-
-const getRiskColor = (r: string) =>
-  r === 'Low'
-    ? 'bg-green-50 text-green-700 border-green-200'
-    : r === 'Medium'
-    ? 'bg-yellow-50 text-yellow-700 border-yellow-200'
-    : r === 'High'
-    ? 'bg-orange-50 text-orange-700 border-orange-200'
-    : r === 'Unlimited'
-    ? 'bg-red-50 text-red-700 border-red-200'
-    : 'bg-gray-50 text-gray-700 border-gray-200';
-
-export const validateParameters = (d: OptionsData) => {
-  if (d.strikePrice <= 0) throw new Error('Strike price must be positive');
-  if (d.currentPrice <= 0) throw new Error('Current price must be positive');
-  if (d.premium < 0) throw new Error('Premium cannot be negative');
-  if (d.daysToExpiry <= 0) throw new Error('Days to expiry must be positive');
-  if (d.impliedVolatility < 0)
-    throw new Error('Implied volatility cannot be negative');
-  if (d.interestRate < 0)
-    throw new Error('Interest rate cannot be negative');
-  if (d.dividendYield < 0)
-    throw new Error('Dividend yield cannot be negative');
-};
-
-const safePayoffCalculation = (
-  s: number,
-  strategy: OptionsStrategy,
-  o: OptionsData,
-) => {
-  try {
-    return strategy.calculatePayoff(s, o);
-  } catch (e) {
-    console.warn('Payoff calculation error:', e);
-    return 0;
-  }
-};
+export { validateParameters } from '../lib/optionsUtils';
 
 export const InteractiveOptionsChart = () => {
   const [expandedStrategy, setExpandedStrategy] = useState<string | null>(
@@ -314,8 +47,8 @@ export const InteractiveOptionsChart = () => {
     }
   }, [optionsData]);
 
-  const generatePayoffData = useMemo(() => {
-    return (strategy: OptionsStrategy): PayoffPoint[] => {
+  const generatePayoffData = useMemo(
+    () => (strategy: OptionsStrategy): PayoffPoint[] => {
       const data: PayoffPoint[] = [];
       const minPrice = Math.max(0, optionsData.strikePrice - 30);
       const maxPrice = optionsData.strikePrice + 30;
@@ -326,200 +59,9 @@ export const InteractiveOptionsChart = () => {
       }
 
       return data;
-    };
-  }, [optionsData]);
-
-  const ParameterSlider = (
-    label: keyof OptionsData,
-    min: number,
-    max: number,
-    step: number,
-  ) => (
-    <div className="mb-4" key={label}>
-      <div className="flex justify-between mb-1">
-        <span className="text-sm capitalize">
-          {label.replace(/([A-Z])/g, ' $1')}
-        </span>
-        <span className="text-sm">{optionsData[label]}</span>
-      </div>
-      <Tooltip>
-        <TooltipTrigger>
-          <Slider
-            aria-label={label}
-            aria-valuetext={`${optionsData[label]}`}
-            min={min}
-            max={max}
-            step={step}
-            value={[optionsData[label]]}
-            onValueChange={(v) =>
-              setOptionsData((prev) => ({ ...prev, [label]: v[0] }))
-            }
-          />
-        </TooltipTrigger>
-        <TooltipContent>
-          <strong>{parameterTooltips[label].title}</strong>
-          <p>{parameterTooltips[label].content}</p>
-        </TooltipContent>
-      </Tooltip>
-    </div>
+    },
+    [optionsData],
   );
-
-  const StrategyCard = ({ strategy }: { strategy: OptionsStrategy }) => {
-    const isExpanded = expandedStrategy === strategy.id;
-    const payoffData = generatePayoffData(strategy);
-    const maxPayoff = Math.max(...payoffData.map((p) => p.profitLoss));
-    const minPayoff = Math.min(...payoffData.map((p) => p.profitLoss));
-    const [selectedTab, setSelectedTab] =
-      useState<'chart' | 'education'>('chart');
-
-    return (
-      <motion.div
-        layout
-        initial={{ opacity: 0, y: 20 }}
-        animate={{ opacity: 1, y: 0 }}
-        transition={{ duration: 0.3 }}
-      >
-        <Card>
-          <CardHeader
-            onClick={() =>
-              setExpandedStrategy(isExpanded ? null : strategy.id)
-            }
-            className="cursor-pointer"
-          >
-            <CardTitle className="flex items-center gap-2">
-              {strategy.title}
-              <ChevronRight
-                className={`w-4 h-4 transition-transform ${
-                  isExpanded ? 'rotate-90' : ''
-                }`}
-              />
-            </CardTitle>
-            <CardDescription>{strategy.description}</CardDescription>
-            <div className="flex gap-1 mt-2">
-              <Badge className={getCategoryColor(strategy.category)}>
-                {strategy.category}
-              </Badge>
-              <Badge className={getComplexityColor(strategy.complexity)}>
-                {strategy.complexity}
-              </Badge>
-              <Badge className={getRiskColor(strategy.riskLevel)}>
-                {strategy.riskLevel}
-              </Badge>
-            </div>
-          </CardHeader>
-          <AnimatePresence>
-            {isExpanded && (
-              <motion.div
-                initial={{ height: 0, opacity: 0 }}
-                animate={{ height: 'auto', opacity: 1 }}
-                exit={{ height: 0, opacity: 0 }}
-                transition={{ duration: 0.3 }}
-              >
-                <CardContent>
-                  <Tabs
-                    value={selectedTab}
-                    onValueChange={(v) =>
-                      setSelectedTab(v as 'chart' | 'education')
-                    }
-                  >
-                    <TabsList>
-                      <TabsTrigger value="chart">Chart</TabsTrigger>
-                      <TabsTrigger value="education">Education</TabsTrigger>
-                    </TabsList>
-                    <TabsContent value="chart">
-                      <div className="w-full h-96">
-                        <ResponsiveContainer width="100%" height="100%">
-                          <LineChart data={payoffData}>
-                            <CartesianGrid
-                              strokeDasharray="3 3"
-                              className="opacity-30"
-                            />
-                            <XAxis
-                              dataKey="stockPrice"
-                              label={{
-                                value: 'Stock Price ($)',
-                                position: 'insideBottom',
-                                offset: -5,
-                              }}
-                            />
-                            <YAxis
-                              label={{
-                                value: 'Profit/Loss ($)',
-                                angle: -90,
-                                position: 'insideLeft',
-                              }}
-                            />
-                            <ReferenceArea
-                              y1={0}
-                              y2={maxPayoff}
-                              fill="rgba(34,197,94,0.1)"
-                            />
-                            <ReferenceArea
-                              y1={minPayoff}
-                              y2={0}
-                              fill="rgba(239,68,68,0.1)"
-                            />
-                            <Line
-                              type="monotone"
-                              dataKey="profitLoss"
-                              stroke="#2563eb"
-                              strokeWidth={3}
-                              dot={false}
-                            />
-                            <ReferenceLine
-                              y={0}
-                              stroke="#374151"
-                              strokeDasharray="2 2"
-                              strokeWidth={2}
-                            />
-                            <ReferenceLine
-                              x={optionsData.currentPrice}
-                              stroke="#2563eb"
-                              strokeDasharray="4 4"
-                              strokeWidth={2}
-                            />
-                            <RechartsTooltip
-                              content={({ active, payload, label }) => {
-                                if (active && payload && payload.length) {
-                                  return (
-                                    <div className="p-2 bg-white border rounded text-sm">
-                                      <p>Stock: {label}</p>
-                                      <p>P/L: {payload[0].value}</p>
-                                      <p>
-                                        {payload[0].value > 0
-                                          ? 'Above the breakeven price the strategy yields a profit'
-                                          : 'Below the breakeven price the strategy loses'}
-                                      </p>
-                                    </div>
-                                  );
-                                }
-                                return null;
-                              }}
-                            />
-                          </LineChart>
-                        </ResponsiveContainer>
-                      </div>
-                    </TabsContent>
-                    <TabsContent value="education">
-                      <StrategyVisualizer
-                        legs={strategy.legs}
-                        optionsData={optionsData}
-                      />
-                      <ul className="list-disc pl-5 mt-2 text-sm">
-                        {strategy.whenToUse.map((w) => (
-                          <li key={w}>{w}</li>
-                        ))}
-                      </ul>
-                    </TabsContent>
-                  </Tabs>
-                </CardContent>
-              </motion.div>
-            )}
-          </AnimatePresence>
-        </Card>
-      </motion.div>
-    );
-  };
 
   return (
     <div className="w-full max-w-7xl mx-auto p-6 space-y-6">
@@ -540,32 +82,11 @@ export const InteractiveOptionsChart = () => {
           </Card>
         </div>
       )}
-      <TooltipProvider>
-        <Card className="mb-6">
-          <CardHeader>
-            <CardTitle className="flex items-center gap-2">
-              <Activity className="w-5 h-5" />
-              Market Parameters & Option Pricing Inputs
-            </CardTitle>
-          </CardHeader>
-          <CardContent>
-            {ParameterSlider('currentPrice', 50, 200, 1)}
-            {ParameterSlider('strikePrice', 50, 200, 5)}
-            {ParameterSlider('premium', 0.5, 30, 0.25)}
-            {ParameterSlider('daysToExpiry', 1, 365, 1)}
-            {ParameterSlider('impliedVolatility', 5, 100, 1)}
-            {ParameterSlider('interestRate', 0, 10, 0.1)}
-            {ParameterSlider('dividendYield', 0, 5, 0.1)}
-            <div className="flex gap-2 mt-4">
-              {Object.entries(quickPresets).map(([k, v]) => (
-                <Button key={k} onClick={() => handlePreset(v)}>
-                  {k}
-                </Button>
-              ))}
-            </div>
-          </CardContent>
-        </Card>
-      </TooltipProvider>
+      <OptionsParameters
+        optionsData={optionsData}
+        onChange={setOptionsData}
+        onPreset={handlePreset}
+      />
       <Card className="mb-6">
         <CardHeader>
           <CardTitle className="flex items-center gap-2">
@@ -579,7 +100,16 @@ export const InteractiveOptionsChart = () => {
       </Card>
       <div className="grid gap-6 grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
         {optionsStrategies.map((s) => (
-          <StrategyCard key={s.id} strategy={s} />
+          <StrategyCard
+            key={s.id}
+            strategy={s}
+            isExpanded={expandedStrategy === s.id}
+            onToggle={(id) =>
+              setExpandedStrategy(expandedStrategy === id ? null : id)
+            }
+            optionsData={optionsData}
+            generatePayoffData={generatePayoffData}
+          />
         ))}
       </div>
       <p className="text-xs text-center text-muted-foreground mt-8">
